uniform redux structure
