+++ src/client/redux/authentication/actions.js
@@ -1,0 +1,64 @@
+import * as type from './types'
+import { API_URL } from '../../helpers/constants'
+
+export const requestLogin = (username, password) => {
+  return {
+    type: type.LOGIN_USER_REQUEST,
+    username,
+    password
+  }
+}
+
+export const requestLoginSuccess = (payload) => {
+  return {
+    type: type.LOGIN_USER_SUCCESS,
+    payload
+  }
+}
+
+const requestLoginError = (message) => {
+  return {
+    type: type.LOGIN_USER_FAILURE,
+    error: message
+  }
+}
+
+// set token on local storage
+const finishAuthentication = (token) => {
+  return localStorage.setItem('token', token)
+}
+
+export const getToken = () => localStorage.getItem('token')
+
+export const authEpic = (action$) => {
+  return (
+    action$.ofType(type.LOGIN_USER_REQUEST)
+      .mergeMap(action => {
+        let user = action.username
+        let password = action.password
+        return fetch(
+          `${API_URL}/${'users/authenticate'}`,
+          {
+            method: 'POST',
+            body: JSON.stringify({user, password}),
+            headers: {
+              'Accept': 'application/json',
+              'Content-Type': 'application/json'
+            }
+          }
+        )
+          .then(response => {
+            return response.json()
+              .then(res => {
+                if (!res.token) {
+                  return requestLoginError('Credentials are wrong')
+                }
+                finishAuthentication(res.token)
+                return requestLoginSuccess()
+              })
+              .catch(error => requestLoginError(error))
+          })
+          .catch(error => requestLoginError(error.message))
+      })
+  )
+}
