+++ src/client/redux/authentication/types.js
@@ -1,0 +1,6 @@
+// action types for authentication
+
+export const LOGIN_USER_REQUEST = 'LOGIN_USER_REQUEST'
+export const LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS'
+export const LOGIN_USER_FAILURE = 'LOGIN_USER_FAILURE'
+export const LOGOUT_USER = 'LOGOUT_USER'
