+++ src/client/redux/greeting/types.js
@@ -1,0 +1,5 @@
+// action types for greeting
+
+export const GET_GREETING = 'GREETING_REQUEST'
+export const GET_GREETING_ERROR = 'GREETING_ERROR'
+export const GET_GREETING_SUCCESS = 'GREETING_SUCCESS'
